+++ federation/mq.test.ts
@@ -1,0 +1,52 @@
+import { Temporal } from "@js-temporal/polyfill";
+import { assertEquals, assertGreater } from "@std/assert";
+import { delay } from "@std/async/delay";
+import { InProcessMessageQueue } from "./mq.ts";
+
+Deno.test("InProcessMessageQueue", async (t) => {
+  const mq = new InProcessMessageQueue();
+
+  const messages: string[] = [];
+  mq.listen((message: string) => {
+    messages.push(message);
+  });
+
+  await t.step("enqueue()", async () => {
+    await mq.enqueue("Hello, world!");
+  });
+
+  await waitFor(() => messages.length > 0, 15_000);
+
+  await t.step("listen()", () => {
+    assertEquals(messages, ["Hello, world!"]);
+  });
+
+  let started = 0;
+  await t.step("enqueue() with delay", async () => {
+    started = Date.now();
+    await mq.enqueue(
+      "Delayed message",
+      { delay: Temporal.Duration.from({ seconds: 3 }) },
+    );
+  });
+
+  await waitFor(() => messages.length > 1, 15_000);
+
+  await t.step("listen() with delay", () => {
+    assertEquals(messages, ["Hello, world!", "Delayed message"]);
+    assertGreater(Date.now() - started, 3_000);
+  });
+});
+
+async function waitFor(
+  predicate: () => boolean,
+  timeoutMs: number,
+): Promise<void> {
+  const started = Date.now();
+  while (!predicate()) {
+    await delay(500);
+    if (Date.now() - started > timeoutMs) {
+      throw new Error("Timeout");
+    }
+  }
+}