+++ httpsig/key.ts
@@ -22,6 +22,7 @@ export function validateCryptoKey(
         "More algorithms will be added in the future!",
     );
   }
+  // @ts-ignore TS2304
   const algorithm = key.algorithm as unknown as RsaHashedKeyAlgorithm;
   if (algorithm.hash.name != "SHA-256") {
     throw new TypeError(
@@ -34,6 +35,7 @@ export function validateCryptoKey(
 /**
  * Generates a key pair which is appropriate for Fedify.
  * @returns The generated key pair.
+ * @since 0.3.0
  */
 export function generateCryptoKeyPair(): Promise<CryptoKeyPair> {
   return crypto.subtle.generateKey(
@@ -54,6 +56,7 @@ export function generateCryptoKeyPair(): Promise<CryptoKeyPair> {
  * @returns The exported key in JWK format.  The key is suitable for
  *          serialization and storage.
  * @throws {TypeError} If the key is invalid or unsupported.
+ * @since 0.3.0
  */
 export async function exportJwk(key: CryptoKey): Promise<JsonWebKey> {
   validateCryptoKey(key);
@@ -66,6 +69,7 @@ export async function exportJwk(key: CryptoKey): Promise<JsonWebKey> {
  * @param type Which type of key to import, either `"public"`" or `"private"`".
  * @returns The imported key.
  * @throws {TypeError} If the key is invalid or unsupported.
+ * @since 0.3.0
  */
 export async function importJwk(
   jwk: JsonWebKey,